use the same response data class for both dukan details and dukan by status endpoints
