+++ dukan/src/main/kotlin/net/thechance/dukan/api/controller/DukanAdminController.kt
@@ -1,12 +1,14 @@
 package net.thechance.dukan.api.controller
 
-import net.thechance.dukan.api.dto.dukan.DukanAdminDetailsResponse
+import net.thechance.dukan.api.dto.dukan.DukanAdminResponse
+import net.thechance.dukan.api.dto.dukan.UpdateDukanStatusRequest
 import net.thechance.dukan.api.dto.product.DukanProductAdminResponse
 import net.thechance.dukan.api.dto.shelf.DukanShelfResponse
 import net.thechance.dukan.api.mapper.dukan.toAdminResponse
 import net.thechance.dukan.api.mapper.product.toAdminResponse
 import net.thechance.dukan.api.mapper.shelf.toResponse
 import net.thechance.dukan.api.utils.EndPoints.DUKAN_PATH
+import net.thechance.dukan.entity.Dukan
 import net.thechance.dukan.entity.DukanProduct
 import net.thechance.dukan.entity.DukanShelf
 import net.thechance.dukan.service.DukanProductService
@@ -19,8 +21,11 @@ import org.springframework.data.domain.Sort
 import org.springframework.data.web.PageableDefault
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PatchMapping
 import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 import java.util.UUID
 
@@ -34,7 +39,7 @@ class DukanAdminController(
 ) {
 
     @GetMapping("/{dukanId}")
-    fun getDukanDetails(@PathVariable("dukanId") dukanId: UUID): ResponseEntity<DukanAdminDetailsResponse> {
+    fun getDukanDetails(@PathVariable("dukanId") dukanId: UUID): ResponseEntity<DukanAdminResponse> {
         val dukan = dukanService.getDukanDetailsById(dukanId)
         val language = LocaleContextHolder.getLocale().language
         val response = dukan.toAdminResponse(language)
@@ -61,5 +66,36 @@ class DukanAdminController(
         val productsPage: Page<DukanProduct> = dukanProductService.getProductsByShelf(shelfId, pageable)
         val response = productsPage.map { product -> product.toAdminResponse() }
         return ResponseEntity.ok(response)
+    }
+
+    @GetMapping
+    fun getDukansByStatusAndSearchQuery(
+        @RequestParam(required = false) query: String = "",
+        @RequestParam status: Dukan.Status,
+        pageable: Pageable
+    ): ResponseEntity<Page<DukanAdminResponse>> {
+        val dukans = dukanService.getDukansByStatusAndQuery(
+            query = query,
+            status = status,
+            pageable = pageable
+        )
+        val language = LocaleContextHolder.getLocale().language
+
+        val response = dukans.map { it.toAdminResponse(language) }
+        return ResponseEntity.ok(response)
+    }
+
+    @PatchMapping("{dukanId}/status")
+    fun updateDukanStatus(
+        @PathVariable("dukanId") dukanId: UUID,
+        @RequestBody updateStatusRequest: UpdateDukanStatusRequest
+    ): ResponseEntity<Unit> {
+        dukanService.updateDukanStatus(
+            dukanId = dukanId,
+            status = updateStatusRequest.status,
+            reason = updateStatusRequest.reason
+        )
+
+        return ResponseEntity.ok().build()
     }
 }
