+++ dukan/src/main/kotlin/net/thechance/dukan/api/mapper/dukan/DukanAdminMapper.kt
@@ -1,10 +1,11 @@
 package net.thechance.dukan.api.mapper.dukan
 
-import net.thechance.dukan.api.dto.dukan.AdminDukanResponse
+import net.thechance.dukan.api.dto.dukan.DukanAdminResponse
+import net.thechance.dukan.api.mapper.category.toDto
 import net.thechance.dukan.entity.Dukan
 
-fun Dukan.toAdminResponse(): AdminDukanResponse {
-    return AdminDukanResponse(
+fun Dukan.toAdminResponse(language: String): DukanAdminResponse {
+    return DukanAdminResponse(
         id = id,
         name = name,
         imageUrl = imageUrl,
@@ -13,6 +14,9 @@ fun Dukan.toAdminResponse(): AdminDukanResponse {
         longitude = longitude,
         status = status,
         activationStatus = activationStatus,
-        createdAt = createdAt
+        createdAt = createdAt,
+        color = color,
+        style = style,
+        categories = categories.map{it.toDto(language)},
     )
 }
