+++ dukan/src/main/kotlin/net/thechance/dukan/api/mapper/dukan/DukanDetailsMapper.kt
@@ -1,6 +1,5 @@
 package net.thechance.dukan.api.mapper.dukan
 
-import net.thechance.dukan.api.dto.dukan.DukanAdminDetailsResponse
 import net.thechance.dukan.api.dto.dukan.DukanDetailsResponse
 import net.thechance.dukan.api.mapper.category.toDto
 import net.thechance.dukan.entity.Dukan
@@ -17,19 +16,5 @@ fun Dukan.toResponse(isFavorite: Boolean): DukanDetailsResponse{
         color = color,
         style =style,
         isFavorite = isFavorite
-    )
-}
-
-fun Dukan.toAdminResponse(language: String): DukanAdminDetailsResponse {
-    return DukanAdminDetailsResponse(
-        id = id,
-        name = name,
-        imageUrl = imageUrl.orEmpty(),
-        address = address,
-        latitude = latitude,
-        longitude = longitude,
-        color = color,
-        style = style,
-        categories = categories.map { it.toDto(language) }
     )
 }
